+++ Thunderstore/CHANGELOG.md
@@ -1,7 +1,12 @@
 ## 0.1.0
-
 - Publish
 
 ## 0.1.1
-
 - Fix unnecessary dependency error
+
+## 0.2.0
+- Add Randomly Lunar configurable property
+- Add Conversion Selection Type configurable property
+- Adjust default weights for a larger difference and a higher focus on void items
+- Nerf default bomb generation frequency (1 -> 0.5)
+- Added debug cheats config option for testing purposes